+++ source/TestCommon/source/DurableFunctionApp.TestCommon/DurableTask/DurableClientExtensions.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Core.TestCommon;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
+using Newtonsoft.Json;
 
 namespace Energinet.DataHub.Core.DurableFunctionApp.TestCommon.DurableTask;
 
@@ -71,6 +72,38 @@ public static class DurableClientExtensions
     }
 
     /// <summary>
+    /// Search for an orchestration that is either running or completed,
+    /// and which was started at, or later, than given <paramref name="createdTimeFrom"/>.
+    ///
+    /// If more than one orchestration exists an exception is thrown.
+    /// </summary>
+    public static async Task<DurableOrchestrationStatus> WaitForOrchestationStartedAsync(
+        this IDurableClient client,
+        DateTime createdTimeFrom,
+        TimeSpan? waitTimeLimit = null)
+    {
+        IReadOnlyCollection<DurableOrchestrationStatus> orchestrationInstances = [];
+        var orchestrationFound = await Awaiter.TryWaitUntilConditionAsync(
+            async () =>
+            {
+                var orchestrationsQueryResult = await GetOrchestrationInstancesAsync(client, createdTimeFrom).ConfigureAwait(false);
+                orchestrationInstances = orchestrationsQueryResult.DurableOrchestrationState.ToList();
+
+                return orchestrationInstances.Count > 0;
+            },
+            timeLimit: waitTimeLimit ?? TimeSpan.FromSeconds(60),
+            delay: TimeSpan.FromSeconds(5)).ConfigureAwait(false);
+
+        if (!orchestrationFound)
+            throw new Exception($"No orchestration instance found within configured wait time limit.");
+
+        if (orchestrationInstances.Count > 1)
+            throw new Exception($"More than one orchestration instance found. Expected 1, but found {orchestrationInstances.Count}");
+
+        return orchestrationInstances.Single();
+    }
+
+    /// <summary>
     /// Wait for orchestration instance to be completed within given <paramref name="waitTimeLimit"/>.
     /// </summary>
     /// <param name="client"></param>
@@ -101,5 +134,52 @@ public static class DurableClientExtensions
         return isCompleted
             ? await client.GetStatusAsync(instanceId, showHistory: true, showHistoryOutput: true).ConfigureAwait(false)
             : throw new Exception($"Orchestration instance '{instanceId}' did not complete within configured wait time limit.");
+    }
+
+    /// <summary>
+    /// Wait for orchestration instance to have a custom status where the given <paramref name="matchFunction"/> returns true
+    /// </summary>
+    /// <exception cref="InvalidCastException">Throws InvalidCastException if CustomStatus property cannot be parsed to given type</exception>
+    public static async Task<DurableOrchestrationStatus> WaitForCustomStatusAsync<TCustomStatus>(
+        this IDurableClient client,
+        string instanceId,
+        Func<TCustomStatus, bool> matchFunction,
+        TimeSpan? waitTimeLimit = null)
+    {
+        var matchesCustomStatus = await Awaiter.TryWaitUntilConditionAsync(
+            async () =>
+            {
+                // Do not retrieve history here as it could be expensive
+                var orchestrationStatus = await client.GetStatusAsync(instanceId).ConfigureAwait(false);
+
+                var customStatus = orchestrationStatus.CustomStatus.ToObject<TCustomStatus>()
+                    ?? throw new InvalidCastException($"Cannot cast CustomStatus to {typeof(TCustomStatus).Name}");
+
+                var isMatch = matchFunction(customStatus);
+
+                return isMatch;
+            },
+            waitTimeLimit ?? TimeSpan.FromSeconds(30),
+            delay: TimeSpan.FromSeconds(5)).ConfigureAwait(false);
+
+        var actualStatus = await client.GetStatusAsync(instanceId, showHistory: true, showHistoryOutput: true).ConfigureAwait(false);
+        return matchesCustomStatus
+            ? actualStatus
+            : throw new Exception($"Orchestration instance '{instanceId}' did not match custom status within configured wait time limit. Actual status: {actualStatus.CustomStatus.ToString(Formatting.Indented)}");
+    }
+
+    private static async Task<OrchestrationStatusQueryResult> GetOrchestrationInstancesAsync(IDurableClient client, DateTime createdTimeFrom)
+    {
+        var filter = new OrchestrationStatusQueryCondition()
+        {
+            CreatedTimeFrom = createdTimeFrom,
+            RuntimeStatus =
+            [
+                OrchestrationRuntimeStatus.Running,
+                OrchestrationRuntimeStatus.Completed,
+            ],
+        };
+        var queryResult = await client.ListInstancesAsync(filter, CancellationToken.None).ConfigureAwait(false);
+        return queryResult;
     }
 }